Make TypeWidgetLibrary interface and impl
